move test to base field test, working on simd test
